+++ CHANGELOG.md
@@ -11,11 +11,16 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
-## v3.0.0-rc.197 (2021-??-??)
+## v3.0.0-rc.197 (2021-06-07)
 
 #### :boom: Breaking Change
 
 * Removed initialization of `iProgress` modifier event listeners `iData`
+
+#### :bug: Bug Fix
+
+* Fixed an issue with `bBottomSlide` not being able to open from the pull
+* Fixed an issue with `bBottomSlide` not being able to initialize `i-history` if the component was opened via swipe
 
 #### :memo: Documentation